pyproject.toml
[project]
name = "my-ml-project"
version = "1.0.0"
description = "My ML project template"
requires-python = "==3.12.9"
readme = "README.md"
authors = [
    { name = "KMYU", email = "sguys99@gmail.com" } # Optional
]
maintainers = [
    { name = "KMYU", email = "sguys99@gmail.com" } # Optional
]

# Project runtime dependencies migrated from requirements.txt for uv sync
dependencies = [
    "amazon-transcribe==0.6.4",
    "apscheduler==3.11.0",
    "azure-cognitiveservices-speech==1.45.0",
    "azure-identity==1.24.0",
    "beautifulsoup4==4.13.5",
    "boto3==1.40.21",
    "dependency-injector==4.48.1",
    "email-validator==2.3.0",
    "fastapi==0.116.1",
    "fastapi-problem==0.11.5",
    "fastapi-sso==0.18.0",
    "google-cloud-aiplatform==1.111.0",
    "jupyter==1.0.0",
    "langchain==0.3.27",
    "langchain-aws==0.2.31",
    "langchain-community==0.3.29",
    "langchain-litellm==0.2.2",
    "langgraph==0.6.6",
    "lightning==2.5.5",
    "litellm==1.76.3",
    "matplotlib==3.10.6",
    "motor==3.7.1",
    "numpy==2.3.3",
    "oauthlib==3.3.1",
    "openpyxl==3.1.2",
    "opensearch-py==3.0.0",
    "opentelemetry-exporter-otlp==1.36.0",
    "opentelemetry-exporter-prometheus==0.57b0",
    "opentelemetry-instrumentation-asyncio==0.57b0",
    "opentelemetry-instrumentation-fastapi==0.57b0",
    "opentelemetry-instrumentation-httpx==0.57b0",
    "opentelemetry-instrumentation-pymongo==0.57b0",
    "opentelemetry-instrumentation-redis==0.57b0",
    "opentelemetry-instrumentation-system-metrics==0.57b0",
    "opentelemetry-instrumentation-threading==0.57b0",
    "optuna==3.2.0",
    "pandas==2.3.3",
    "pg8000==1.31.4",
    "polars==0.20.19",
    "prometheus-client==0.21.1",
    "pyarrow==21.0.0",
    "pydantic==2.11.7",
    "pymongo==4.14.1",
    "python-dotenv==1.1.1",
    "python-json-logger==3.3.0",
    "python-multipart==0.0.20",
    "pyvespa==0.61.0",
    "pyyaml==6.0.2",
    "redis==6.4.0",
    "scikit-learn==1.7.2",
    "seaborn==0.12.2",
    "shap==0.42.1",
    "slack-sdk==3.36.0",
    "sqlalchemy==2.0.43",
    "statsmodels==0.14.0",
    "streamlit==1.50.0",
    "streamlit-aggrid==1.1.9",
    "streamlit-modal==0.1.2",
    "streamlit-option-menu==0.3.12",
    "torch==2.8.0",
    "torchinfo==1.8.0",
    "traceloop-sdk==0.46.2",
    "tqdm==4.65.0",
    "tzlocal==5.3.1",
    "uuid7==0.1.0",
    "xgboost==1.7.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/my_ml"]

[tool.ruff]
line-length = 105
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["my_ml"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[project.optional-dependencies]
dev = [
    "ruff >= 0.8.0",
    "pre-commit >= 3.3.3",
]
